Fix companies not shown on localhost

+++ lib/models/company.dart
@@ -8,7 +8,7 @@ class CompanyModel {
   String? country;
   String? exchange;
   String? industry;
-  int? marketcap;
+  double? marketcap;
   String? companyname;
   String? exchangeshortname;
   String? isactivelytrading;
